Creating login register

+++ Codigo/Cycling-api/src/main/java/com/CyclingConnect/cyclingconnect/repositories/UserRepository.java
@@ -3,7 +3,6 @@ package com.CyclingConnect.cyclingconnect.repositories;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Repository;
 
@@ -14,7 +13,7 @@ import java.util.List;
  * Repositório para operações relacionadas a usuários.
  */
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     /**
      * Busca um usuário pelo login.
